requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
lightgbm==4.3.0
tensorflow==2.18.0
tensorflow-metal==1.2.0
flask==3.0.3
requests==2.32.3
implicit==0.7.2
faiss-cpu==1.8.0
scipy==1.11.4
packaging==24.1
